+++ js_api/src/service/parachain.ts
@@ -188,33 +188,35 @@ async function _queryAuctionInfo(api: ApiPromise) {
 }
 
 /**
+ * query parachain module overview.
+ */
+async function queryParasOverview(api: ApiPromise) {
+  const [paras, bestNumber] = await Promise.all([api.query.paras.parachains(), api.derive.chain.bestNumber()]);
+
+  const length = api.consts.slots.leasePeriod as BlockNumber;
+  const startNumber = bestNumber.sub((api.consts.slots.leaseOffset as BlockNumber) || BN_ZERO);
+
+  return {
+    parasCount: (paras as any).length,
+    currentLease: startNumber.div(length).toNumber(),
+    leaseLength: length.toNumber(),
+    leaseProgress: startNumber.mod(length).toNumber(),
+  };
+}
+
+/**
  * query crowd loan contributions of an account.
  *
  * @param {String} paraId
  * @param {String} pubKey
  */
 async function queryUserContributions(api: ApiPromise, paraId: string, pubKey: string) {
-  const fund = await api.query.crowdloan.funds(paraId) as any;
-  const childKey = _createChildKey(fund.unwrap().trieIndex);
-  const value = await api.rpc.childstate.getStorage(childKey, pubKey);
-  if (value.isSome) {
-    return api.createType('(Balance, Vec<u8>)' as any, value.unwrap()).toJSON()[0].toString();
-  }
-  return '0';
-}
-
-function _createChildKey (trieIndex: TrieIndex): string {
-  return u8aToHex(
-    u8aConcat(
-      ':child_storage:default:',
-      blake2AsU8a(
-        u8aConcat('crowdloan', trieIndex.toU8a())
-      )
-    )
-  );
+  const res = await api.derive.crowdloan.ownContributions(paraId, [pubKey]);
+  return res[pubKey];
 }
 
 export default {
   queryAuctionWithWinners,
+  queryParasOverview,
   queryUserContributions
 };