+++ lib/api/apiParachain.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 
 import 'package:polkawallet_sdk/api/api.dart';
 import 'package:polkawallet_sdk/api/types/parachain/auctionData.dart';
+import 'package:polkawallet_sdk/api/types/parachain/parasOverviewData.dart';
 import 'package:polkawallet_sdk/service/parachain.dart';
 
 class ApiParachain {
@@ -10,10 +11,14 @@ class ApiParachain {
   final PolkawalletApi apiRoot;
   final ServiceParachain service;
 
+  Future<ParasOverviewData> queryParasOverview() async {
+    final res = await service.queryParasOverview();
+    return ParasOverviewData.fromJson(res ?? {});
+  }
+
   Future<AuctionData> queryAuctionWithWinners() async {
-    final res = await (service.queryAuctionWithWinners()
-        as FutureOr<Map<dynamic, dynamic>>);
-    return AuctionData.fromJson(res);
+    final res = await service.queryAuctionWithWinners();
+    return AuctionData.fromJson(res!);
   }
 
   Future<List<String>> queryUserContributions(